\name{GelitoMain}

\alias{GelitoMain}

\title{
Main Class for Gelito.
}

\description{
Main Class for Gelito. It is instanced into Gelito.
}

\usage{
GelitoMain()
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\note{
For using with Gelito package.
}

\seealso{
Gelito
}

\examples{
gm <- GelitoMain()
}

