\name{getEnvironment}

\alias{getEnvironment}

\title{
Get list with environment attributes.
}

\description{
Get list with environment attributes.
}

\usage{
gm <- GelitoMain()
gm$getEnvironment()
}

\value{
  \item{list}{
with environment parameters
}
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
