\name{getEnvironmentShiny}

\alias{getEnvironmentShiny}

\title{
Return string version for environment. Useful for Shiny.
}

\description{
Return string version for environment. Useful for Shiny..
}

\usage{
gm <- GelitoMain()
gm$getEnvironmentShiny()
}

\value{
show string version of enviroment for Shiny
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
