\name{getEnvironmentString}

\alias{getEnvironmentString}

\title{
Return string version for environment.
}

\description{
Return string version for environment.
}

\usage{
gm <- GelitoMain()
gm$getEnvironmentString()
}

\value{
show string version of enviroment
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
