\name{inputEnvironment}

\alias{inputEnvironment}

\title{
Get environment parameters from console.
}

\description{
Get environment parameters from console. This method fill attribute list.
}

\usage{
gm <- GelitoMain()
gm$inputEnvironment()
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
