\name{readInteger}

\alias{readInteger}

\title{
Read a positive integer.
}

\description{
Read an positive integer greater than 0.
}

\usage{
gm <- GelitoMain()
gm$readInteger(text)
}

\arguments{
  \item{text}{
text for prompt. It can be blank.
}
}

\value{
  \item{integer}{
number
}
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
