\name{setEnvironment}

\alias{setEnvironment}

\title{
Set environment from params.
}

\description{
Set environment from params.
}

\usage{
gm <- GelitoMain()
gm$setEnvironment(id = NA, nf = list(), cc = list(), cn = list())
}

\arguments{
  \item{id}{
ID column number.
}
  \item{nf}{
filter columns numbers.
}
  \item{cc}{
criterions.
}
  \item{cn}{
column numbers.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
