\name{trim}

\alias{trim}

\title{
Trim a string.
}

\description{
Trim a string.
}

\usage{
gm <- GelitoMain()
gm$trim(x)
}

\arguments{
  \item{x}{
String for trimming.
}
}

\value{
  \item{string}{
a trimmed string, leading and trailing.
}
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
