\name{writeCSVFile}

\alias{writeCSVFile}

\title{
Write CSV file with results.
}

\description{
Write CSV file with results in data attribute.
}

\usage{
gm <- GelitoMain()
gm$writeCSVFile(filename = NA)
}

\arguments{
  \item{filename}{
output filename.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
