\name{writeTXTFile}

\alias{writeTXTFile}

\title{
Write TXT file with results.
}

\description{
Write TXT file with results in data attribute (tab delimited).
}

\usage{
gm <- GelitoMain()
gm$writeTXTFile(filename = NA)
}

\arguments{
  \item{filename}{
output filename.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
