\name{writeXLSFile}

\alias{writeXLSFile}

\title{
Write XLS file with results.
}

\description{
Write XLS file with results in data attribute.
}

\usage{
gm <- GelitoMain()
gm$writeXLSFile(filename = NA)
}

\arguments{
  \item{filename}{
output filename.
}
}

\value{
no
}

\author{
Manuel Ramón Gutiérrez Covarrubias
}

\seealso{
GelitoMain
}
